requirements.txt
elevenlabs==1.7.0
icecream==2.1.3
openai==1.43.0
pytest==8.2.2
python-dotenv==1.0.1
pyttsx3==2.90
rich==13.8.0
simpleaudio==1.0.4
SpeechRecognition==3.10.4
